+++ dev_project/project_env/services/__init__.py
@@ -1,5 +1,4 @@
 from .docker_base_image import BaseImageService
-from .platform_sources import PlatformSourcesService
 from .pycharm_configurator import PycharmConfigurator
 from .python_analysis_paths import PythonAnalysisPathsBuilder
 from .vscode_configurator import VscodeConfigurator
@@ -7,7 +6,6 @@ from .vscode_configurator import VscodeConfigurator
 __all__ = [
     "BaseImageService",
     "CiImageBuildService",
-    "PlatformSourcesService",
     "PycharmConfigurator",
     "PythonAnalysisPathsBuilder",
     "VscodeConfigurator",